+++ ThirdPartyAdapters/meta/meta/src/main/java/com/google/ads/mediation/facebook/FacebookRewardedAd.java
@@ -133,14 +133,13 @@ public class FacebookRewardedAd implements MediationRewardedAd, RewardedVideoAdE
   @Override
   public void onError(Ad ad, com.facebook.ads.AdError adError) {
     AdError error = FacebookMediationAdapter.getAdError(adError);
+    Log.w(TAG, error.getMessage());
 
     if (showAdCalled.get()) {
-      Log.w(TAG, error.getMessage());
       if (rewardedAdCallback != null) {
         rewardedAdCallback.onAdFailedToShow(error);
       }
     } else {
-      Log.w(TAG, error.getMessage());
       if (mediationAdLoadCallback != null) {
         mediationAdLoadCallback.onFailure(error);
       }